Add system tests
In test_mpop_kiosk, we check that some JSON-RPC methods return the
expected results.

It requires the Docker Compose services to be running, so that you
have an MPOP Service that is available.

You can change the service host name and port number by passing
command-line arguments to the test suite in Qt Creator.
For example: `10.10.10.10 3334` will configure the test suite to
communicate with the MPOP Service on that host IP and port number.
You can pass only the first argument, so that you keep using the default
port number, like this: `10.10.10.10`.
You can set the command-line arguments to the test suite by clicking on
"Project" and then "Run". Then, populate the "Command line arguments"
text field.